+++ src/lib/components/common/RichTextInput/Youtube/index-manual.ts
@@ -1,2 +1,0 @@
-export { YoutubeManual as default } from './youtube-manual';
-export * from './youtube-manual';